Add default value to select_rows columns array
When selecting rows, it seems like a common usage is to just select all of them. However, sending in the value ["*"] is unintuitive and might not cross peoples mind.

Here I add a default value of ["*"] in the select_rows function so that the columns value can just be omitted to select all coulumns.